creating a basic model for client-server messaging

+++ server/src/settings.h
@@ -1,12 +1,13 @@
 #pragma once
 #include <string>
+#include <cstdint>
 
 namespace ServerDigitalTwin {
 extern const std::string CONFIG_FILE;
 
 struct URL {
   std::string host;
-  long port;
+  uint16_t port;
 };
 
 } // namespace ServerDigitalTwin
